P3/django_back/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
idna==3.3
olefile==0.46
Pillow==9.0.1
platformdirs==2.5.1
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
virtualenv==20.13.2
django-cors-headers==3.11.0